renamed Client to Instance
